+++ src/main/java/fullydynamic/FullyDynamicSubgraphReservoirAlgorithm.java
@@ -21,7 +21,7 @@ public class FullyDynamicSubgraphReservoirAlgorithm implements TopkGraphPatterns
 	NodeMap nodeMap;
 	EdgeHandler utility;
 	SubgraphReservoir<Triplet> reservoir;
-	HashMap<Pattern, Integer> frequentPatterns;
+	HashMap<Pattern, Long> frequentPatterns;
 	int N; // total number of subgraphs
 	int M; // maximum reservoir size
 	int Ncurrent;
@@ -36,7 +36,7 @@ public class FullyDynamicSubgraphReservoirAlgorithm implements TopkGraphPatterns
 		c1=0;
 		Ncurrent = 0 ;
 		c2=0;
-		frequentPatterns = new HashMap<Pattern, Integer>();
+		frequentPatterns = new HashMap<Pattern, Long>();
 	}
 
 	public boolean addEdge(StreamEdge edge) {
@@ -184,17 +184,17 @@ public class FullyDynamicSubgraphReservoirAlgorithm implements TopkGraphPatterns
 	void addFrequentPattern(Triplet t) {
 		ThreeNodeGraphPattern p = new ThreeNodeGraphPattern(t);
 		if(frequentPatterns.containsKey(p)) {
-			int count = frequentPatterns.get(p);
+			long count = frequentPatterns.get(p);
 			frequentPatterns.put(p, count+1);
 		}else {
-			frequentPatterns.put(p, 1);
+			frequentPatterns.put(p, 1l);
 		}
 	}
 	
 	void removeFrequentPattern(Triplet t) {
 		ThreeNodeGraphPattern p = new ThreeNodeGraphPattern(t);
 		if(frequentPatterns.containsKey(p)) {
-			int count = frequentPatterns.get(p);
+			long count = frequentPatterns.get(p);
 			if(count >1)
 				frequentPatterns.put(p, count-1);
 			else 
@@ -202,16 +202,16 @@ public class FullyDynamicSubgraphReservoirAlgorithm implements TopkGraphPatterns
 		}
 	}
 	
-	public HashMap<Pattern, Integer> getFrequentPatterns() {
+	public HashMap<Pattern, Long> getFrequentPatterns() {
 		return this.frequentPatterns;
 	}
 	public void correctEstimates() {
 		double correctFactor = correctFactor();
 		List<Pattern> patterns = new ArrayList<Pattern>(frequentPatterns.keySet());
 		for(Pattern p: patterns) {
-			int count = frequentPatterns.get(p);
+			long count = frequentPatterns.get(p);
 			double value = count*correctFactor;
-			frequentPatterns.put(p, (int)value);
+			frequentPatterns.put(p, (long)value);
 		}
 	}
 	private double correctFactor() {
