+++ src/main/java/incremental/IncrementalExhaustiveCountingFourNode.java
@@ -20,14 +20,14 @@ import utility.QuadripletGenerator;
 public class IncrementalExhaustiveCountingFourNode implements TopkGraphPatterns {
 	NodeMap nodeMap;
 	EdgeHandler utility;
-	HashMap<Pattern, Integer> frequentPatterns;
+	HashMap<Pattern, Long> frequentPatterns;
 	int numSubgraph;
 	QuadripletGenerator subgraphGenerator;
 
 	public IncrementalExhaustiveCountingFourNode() {
 		utility = new EdgeHandler();
 		numSubgraph = 0;
-		frequentPatterns = new HashMap<Pattern, Integer>();
+		frequentPatterns = new HashMap<Pattern, Long>();
 		this.nodeMap = new NodeMap();
 		subgraphGenerator = new QuadripletGenerator();
 	}
@@ -81,17 +81,17 @@ public class IncrementalExhaustiveCountingFourNode implements TopkGraphPatterns 
 	void addFrequentPattern(Quadriplet t) {
 		FourNodeGraphPattern p = new FourNodeGraphPattern(t);
 		if (frequentPatterns.containsKey(p)) {
-			int count = frequentPatterns.get(p);
+			long count = frequentPatterns.get(p);
 			frequentPatterns.put(p, count + 1);
 		} else {
-			frequentPatterns.put(p, 1);
+			frequentPatterns.put(p, 1l);
 		}
 	}
 
 	void removeFrequentPattern(Quadriplet t) {
 		FourNodeGraphPattern p = new FourNodeGraphPattern(t);
 		if (frequentPatterns.containsKey(p)) {
-			int count = frequentPatterns.get(p);
+			long count = frequentPatterns.get(p);
 			if (count > 1)
 				frequentPatterns.put(p, count - 1);
 			else
@@ -99,7 +99,7 @@ public class IncrementalExhaustiveCountingFourNode implements TopkGraphPatterns 
 		}
 	}
 
-	public HashMap<Pattern, Integer> getFrequentPatterns() {
+	public HashMap<Pattern, Long> getFrequentPatterns() {
 		return this.frequentPatterns;
 	}
