+++ src/main/java/incremental/IncrementalSubgraphReservoirAlgorithm.java
@@ -21,7 +21,7 @@ public class IncrementalSubgraphReservoirAlgorithm implements TopkGraphPatterns 
 	NodeMap nodeMap;
 	EdgeHandler utility;
 	SubgraphReservoir<Triplet> reservoir;
-	HashMap<Pattern, Integer> frequentPatterns;
+	HashMap<Pattern, Long> frequentPatterns;
 	
 	int N; // total number of subgraphs
 	int M; // maximum reservoir size
@@ -31,7 +31,7 @@ public class IncrementalSubgraphReservoirAlgorithm implements TopkGraphPatterns 
 		reservoir = new SubgraphReservoir<Triplet>();
 		N = 0;
 		M = size;
-		frequentPatterns = new HashMap<Pattern, Integer>();
+		frequentPatterns = new HashMap<Pattern, Long>();
 		
 	}
 
@@ -133,10 +133,10 @@ public class IncrementalSubgraphReservoirAlgorithm implements TopkGraphPatterns 
 	void addFrequentPattern(Triplet t) {
 		ThreeNodeGraphPattern p = new ThreeNodeGraphPattern(t);
 		if(frequentPatterns.containsKey(p)) {
-			int count = frequentPatterns.get(p);
+			long count = frequentPatterns.get(p);
 			frequentPatterns.put(p, count+1);
 		}else {
-			frequentPatterns.put(p, 1);
+			frequentPatterns.put(p, 1l);
 		}
 		
 
@@ -146,7 +146,7 @@ public class IncrementalSubgraphReservoirAlgorithm implements TopkGraphPatterns 
 	void removeFrequentPattern(Triplet t) {
 		ThreeNodeGraphPattern p = new ThreeNodeGraphPattern(t);
 		if(frequentPatterns.containsKey(p)) {
-			int count = frequentPatterns.get(p);
+			long count = frequentPatterns.get(p);
 			if(count >1)
 				frequentPatterns.put(p, count-1);
 			else 
@@ -157,16 +157,16 @@ public class IncrementalSubgraphReservoirAlgorithm implements TopkGraphPatterns 
 	}
 	
 	
-	public HashMap<Pattern, Integer> getFrequentPatterns() {
+	public HashMap<Pattern, Long> getFrequentPatterns() {
 		return this.frequentPatterns;
 	}
 	public void correctEstimates() {
 		double correctFactor = correctFactor();
 		List<Pattern> patterns = new ArrayList<Pattern>(frequentPatterns.keySet());
 		for(Pattern p: patterns) {
-			int count = frequentPatterns.get(p);
+			long count = frequentPatterns.get(p);
 			double value = count*correctFactor;
-			frequentPatterns.put(p, (int)value);
+			frequentPatterns.put(p, (long)value);
 		}
 	}
 	private double correctFactor() {
