+++ src/main/java/incremental/IncrementalSubgraphReservoirFinalAlgorithmFourNode2.java
@@ -35,7 +35,7 @@ public class IncrementalSubgraphReservoirFinalAlgorithmFourNode2 implements Topk
 	Random rand;
 	QuadripletGenerator subgraphGenerator;
 
-	HashMap<Pattern, Integer> frequentPatterns;
+	HashMap<Pattern, Long> frequentPatterns;
 	int numberSubgraphs; // total number of subgraphs
 	int reservoirSize; // maximum reservoir size
 	int sum;
@@ -49,7 +49,7 @@ public class IncrementalSubgraphReservoirFinalAlgorithmFourNode2 implements Topk
 		reservoir = new AdvancedSubgraphReservoir<Quadriplet>();
 		numberSubgraphs = 0;
 		reservoirSize = size;
-		frequentPatterns = new HashMap<Pattern, Integer>();
+		frequentPatterns = new HashMap<Pattern, Long>();
 		sum = 0;
 		skipRS = new AlgorithmZ(reservoirSize);
 	}
@@ -172,17 +172,17 @@ public class IncrementalSubgraphReservoirFinalAlgorithmFourNode2 implements Topk
 	void addFrequentPattern(Quadriplet t) {
 		FourNodeGraphPattern p = new FourNodeGraphPattern(t);
 		if (frequentPatterns.containsKey(p)) {
-			int count = frequentPatterns.get(p);
+			long count = frequentPatterns.get(p);
 			frequentPatterns.put(p, count + 1);
 		} else {
-			frequentPatterns.put(p, 1);
+			frequentPatterns.put(p, 1l);
 		}
 	}
 
 	void removeFrequentPattern(Quadriplet t) {
 		FourNodeGraphPattern p = new FourNodeGraphPattern(t);
 		if (frequentPatterns.containsKey(p)) {
-			int count = frequentPatterns.get(p);
+			long count = frequentPatterns.get(p);
 			if (count > 1)
 				frequentPatterns.put(p, count - 1);
 			else
@@ -190,7 +190,7 @@ public class IncrementalSubgraphReservoirFinalAlgorithmFourNode2 implements Topk
 		}
 	}
 
-	public HashMap<Pattern, Integer> getFrequentPatterns() {
+	public HashMap<Pattern, Long> getFrequentPatterns() {
 		return this.frequentPatterns;
 	}
 
@@ -198,9 +198,9 @@ public class IncrementalSubgraphReservoirFinalAlgorithmFourNode2 implements Topk
 		double correctFactor = correctFactor();
 		List<Pattern> patterns = new ArrayList<Pattern>(frequentPatterns.keySet());
 		for (Pattern p : patterns) {
-			int count = frequentPatterns.get(p);
+			long count = frequentPatterns.get(p);
 			double value = count * correctFactor;
-			frequentPatterns.put(p, (int) value);
+			frequentPatterns.put(p, (long) value);
 		}
 	}
