+++ src/main/java/topkgraphpattern/TopkGraphPatterns.java
@@ -8,7 +8,7 @@ import input.StreamEdge;
 public interface TopkGraphPatterns {
 	public boolean addEdge(StreamEdge edge);
 	public boolean removeEdge(StreamEdge edge);
-	public HashMap<Pattern, Integer> getFrequentPatterns();
+	public HashMap<Pattern, Long> getFrequentPatterns();
 	public int getNumberofSubgraphs();
 	public int getCurrentReservoirSize();
 	public void correctEstimates();
